REF: Move OrderedMessageHandler into separate module
